SatTraker Beta Version 1
EXPERIMENTAL, USE AT YOUR OWN RISK!  ASCOM compatibility has not been verified against actual hardware.  This program is designed to track earth orbiting satellites using a TLE file and guide those satellites to the center of the view of a telescope using a video viewfinder or guidescope with a compatible video camera.  Please see Astronomy Live on YouTube for details on how to use this software.
https://www.youtube.com/messierhunter