Linux support for LX200: V4L (via OpenCV) and feature tracking works Connecting via using the whole device name (ex: /dev/ttyACM0) works for the telescope.
A couple of lines for initial INDI support. (Hidden for now)
Removed the ASCOM option for non-windows platforms.

Untested:
Actual movement on a scope.
Brightneess tracking (Using my computer, and available LEDs are either too bright or too dim for a contrast.) I suspect it works.

Unable to be tested:
I suspect that LX200 usage will be supported on Macs, I don't have a mac though.